support for adding markers only when visible in GridClusteringStrategy changed clusterId to long because of overflow on high zoom

+++ android-maps-extensions/src/pl/mg6/android/maps/extensions/impl/ClusterMarker.java
@@ -25,7 +25,7 @@ import com.google.android.gms.maps.model.LatLngBounds;
 
 class ClusterMarker implements Marker {
 
-	private int clusterId;
+	private long clusterId;
 
 	private int lastCount = -1;
 
@@ -39,11 +39,11 @@ class ClusterMarker implements Marker {
 		this.strategy = strategy;
 	}
 
-	int getClusterId() {
+	long getClusterId() {
 		return clusterId;
 	}
 
-	void setClusterId(int clusterId) {
+	void setClusterId(long clusterId) {
 		this.clusterId = clusterId;
 	}
